requirements.txt
streamlit>=1.32.0
pillow>=10.0.0
numpy>=1.24.0
pandas>=2.0.0

torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
segment-anything @ git+https://github.com/facebookresearch/segment-anything.git

pvlib>=0.10.4
scipy>=1.10.0
matplotlib>=3.7.0
xarray>=2023.0.0

requests>=2.31.0
meteostat>=1.6.0

fpdf>=1.7.2

python-dateutil>=2.8.2
pytz>=2023.3